-Changed ColorEncoder encodeColorHsb* methods slightly

+++ src/model/ColorEncoder.java
@@ -38,7 +38,7 @@ public class ColorEncoder {
     private int encodeColorHsbLinear(int iterations, int maxIterations){
         float color = 1.0f*iterations/maxIterations;
         color += colorOffset/255.0;
-        return Color.HSBtoRGB((float) color, 1, 1);
+        return Color.HSBtoRGB(color, 1, 1);
     }
 
     private int encodeColorHsbLogarithmic(int iterations, int maxIterations){
